+++ src/main/java/com/solace/samples/jcsmp/samplers/QueueProvisionAndRequestActiveFlowIndication.java
@@ -10,10 +10,10 @@
 
 package com.solace.samples.jcsmp.samplers;
 
-import com.solace.samples.samplers.jcsmp.common.ArgParser;
-import com.solace.samples.samplers.jcsmp.common.SampleApp;
-import com.solace.samples.samplers.jcsmp.common.SampleUtils;
-import com.solace.samples.samplers.jcsmp.common.SessionConfiguration;
+import com.solace.samples.jcsmp.samplers.common.ArgParser;
+import com.solace.samples.jcsmp.samplers.common.SampleApp;
+import com.solace.samples.jcsmp.samplers.common.SampleUtils;
+import com.solace.samples.jcsmp.samplers.common.SessionConfiguration;
 import com.solacesystems.jcsmp.BytesXMLMessage;
 import com.solacesystems.jcsmp.CapabilityType;
 import com.solacesystems.jcsmp.ConsumerFlowProperties;
